Remove outdated tests and related code from the testsuite
- Deleted logtable.go, query.go, tables.go, tagtable.go, testsuite.go, users.go files.
- These files contained tests that are no longer relevant or needed for the current codebase.
- This cleanup helps streamline the testing process and maintain focus on active test cases.
